Move custom_link/2 (and related functions) to Autolink
1. identify function to move over: "custom_link/2"
2. copy-paste, from elixir.ex, then remove it there
3. save, format and mix test
4. got warnings, solved those, identified new function
5. rinse and repeat from 2-4 until no more warnings

Finally, "Autolink." was removed from autolink.ex
erlang.ex was adapted to use the newly exposed function

Tests are as-were (no failures, locally)
There are no compilation warnings, locally
